Improving rendering of border-left
Prior to this commit, the border-left often would render a line that
started above the content.  With this change, I'm aiming to only render
the border to the left of content.

Below is an example of before state for DETAILS and BLOCKQUOTE sections:

```
Content
|
|  Content
|
|  Content
|
|  Content

Content
```

The `|` is the visual border.

The after state with this change is:

```
Content

|  Content
|
|  Content
|
|  Content

Content
```