Working on details and border
I believe
http://takeonrules.com/2020/05/10/upcoming-dungeons-and-dragons-game/ is
my most complicated DETAILS with Marginalia interaction

The restoration of `p, summary { padding-top: 1.4rem; }` ensures that a
post with a MARGINFIGURE first element has a properly padded image.
When it was margin-top, the image butted up against the list of tags.

The commit that introduced this was somewhat lazy.  I thought I could
make the change without impact.  The reason for the initial change was
to ensure that a paragraph rendered within a DETAILS would have a
contiguous left border; That is resolved by switching from padding to margin.